+++ includes/class-comment.php
@@ -647,8 +647,13 @@ class Comment {
 				'icon'           => '♻️',
 				'class'          => 'p-repost',
 				'type'           => 'repost',
+				'collection'     => 'reposts',
 				'activity_types' => array( 'announce' ),
 				'excerpt'        => __( '&hellip; reposted this!', 'activitypub' ),
+				/* translators: %d: Number of reposts */
+				'count_single'   => _x( '%d repost', 'number of reposts', 'activitypub' ),
+				/* translators: %d: Number of reposts */
+				'count_plural'   => _x( '%d reposts', 'number of reposts', 'activitypub' ),
 			)
 		);
 
@@ -661,8 +666,13 @@ class Comment {
 				'icon'           => '👍',
 				'class'          => 'p-like',
 				'type'           => 'like',
+				'collection'     => 'likes',
 				'activity_types' => array( 'like' ),
 				'excerpt'        => __( '&hellip; liked this!', 'activitypub' ),
+				/* translators: %d: Number of likes */
+				'count_single'   => _x( '%d like', 'number of likes', 'activitypub' ),
+				/* translators: %d: Number of likes */
+				'count_plural'   => _x( '%d likes', 'number of likes', 'activitypub' ),
 			)
 		);
 	}