+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## Unreleased
+
+### Improved
+* Reactions block: Remove the `wp-block-editor` dependency for frontend views 
+
+## [4.5.0] - 2024-12-17
 
 ### Added
 
+* Reactions block to display likes and reposts
 * `icon` support for `Audio` and `Video` attachments
 * Send "new follower" emails
 * Send "direct message" emails
@@ -17,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improved
 
+* Likes and Reposts enabled by default
 * Email templates for Likes and Reposts
 * Improve Interactions moderation
 * Compatibility with Akismet
@@ -523,7 +530,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Normalize attributes that can have mixed value types
 
-## [1.3.0] 2023-12-05
+## [1.3.0] - 2023-12-05
 
 ### Added
 
@@ -1148,8 +1155,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * initial
 
-[Unreleased]: https://github.com/Automattic/wordpress-activitypub/compare/4.4.0...trunk
+[Unreleased]: https://github.com/Automattic/wordpress-activitypub/compare/4.5.0...trunk
 <!-- Add new release below and update "Unreleased" link -->
+[4.5.0]: https://github.com/Automattic/wordpress-activitypub/compare/4.4.0...4.5.0
 [4.4.0]: https://github.com/Automattic/wordpress-activitypub/compare/4.3.0...4.4.0
 [4.3.0]: https://github.com/Automattic/wordpress-activitypub/compare/4.2.1...4.3.0
 [4.2.1]: https://github.com/Automattic/wordpress-activitypub/compare/4.2.0...4.2.1